# perf(core): use native addEventListener for faster rendering.